Move to graphql folder
